+++ src/app/layout.tsx
@@ -123,7 +123,7 @@ export default function RootLayout({ children }: { children: ReactNode }) {
             <AppSidebar />
             <div
               id="layout-main-content-container"
-              className="flex h-screenish w-full overflow-x-hidden mt-12 mb-4 mr-4 border-t border-l border-border rounded-lg bg-background"
+              className="flex h-[calc(100vh-44px)] w-full overflow-x-hidden mt-11 bg-background"
             >
               {children}
             </div>